+++ CHANGELOG.md
@@ -83,6 +83,7 @@ CHANGELOG
     new: `public function queryAndReturnResult(string $query, ?array $variables = null, array $opts = []): ExecutionResult`
 
 ### Added
+- Command to make an exection middleware [\#772 / mfn](https://github.com/rebing/graphql-laravel/pull/772)
 - The primary execution of the GraphQL request is now piped through middlewares [\#762 / crissi and mfn](https://github.com/rebing/graphql-laravel/pull/762)\
   This allows greater flexibility for enabling/disabling certain functionality
   as well as bringing in new features without having to open up the library.